+++ src/.config/nuxt.ts
@@ -119,6 +119,18 @@ export default defineNuxtConfig({
       plugins: [vue()],
     },
   },
+  routeRules: {
+    '/api/auth-proxy': {
+      security: {
+        xssValidator: false, // TipTap's HTML is stored unescaped (is escaped when displayed) so api requests would trigger the xss protection on forward authentication (https://github.com/maevsi/maevsi/issues/1603)
+      },
+    },
+    '/api/ical': {
+      security: {
+        xssValidator: false, // TipTap's HTML is stored unescaped (is escaped when displayed) so api requests would trigger the xss protection here (https://github.com/maevsi/maevsi/issues/1603)
+      },
+    },
+  },
   runtimeConfig: {
     public: {
       i18n: {
@@ -149,6 +161,9 @@ export default defineNuxtConfig({
             sampleRate: 0.0,
           },
         },
+      },
+      security: {
+        isRateLimiterDisabled: true, // TODO: disable once api requests are optimized (https://github.com/maevsi/maevsi/issues/1654)
       },
       site: {
         url: SITE_URL,