+++ tests/unit/ical.test.ts
@@ -47,11 +47,12 @@ DTSTART:19700101T000000Z\r
 DTEND:19700101T000000Z\r
 SUMMARY:name\r
 LOCATION:location\r
-DESCRIPTION:https://maevsi.test/events/authorUsername/slug\\n\\ndescription\r
-X-ALT-DESC;FMTTYPE=text/html:https://maevsi.test/events/authorUsername/slu\r
- g\\n<p>description</p>\r
+DESCRIPTION:https://maevsi.test/event/view/authorUsername/slug\\n\\ndescript\r
+ ion\r
+X-ALT-DESC;FMTTYPE=text/html:https://maevsi.test/event/view/authorUsername\r
+ /slug\\n<p>description</p>\r
 ORGANIZER;CN="authorUsername":mailto:authorUsername@maevsi.test\r
-URL;VALUE=URI:https://maevsi.test/events/authorUsername/slug\r
+URL;VALUE=URI:https://maevsi.test/event/view/authorUsername/slug\r
 STATUS:CONFIRMED\r
 END:VEVENT\r
 END:VCALENDAR`,